Moved developer components into Developer folder
