+++ src/components/Developer/NetworkLogging.js
@@ -1,10 +1,9 @@
 // @flow
 
+import ViewWrapper from "components/SharedComponents/ViewWrapper";
 import type { Node } from "react";
 import React from "react";
 import NetworkLogger from "react-native-network-logger";
-
-import ViewWrapper from "./SharedComponents/ViewWrapper";
 
 // TODO: remove this from production build
 const NetworkLogging = (): Node => (
